Created Containers for objects.
To clear up the scene, I created containers for the static objects, such as the castles in the background and the "enemy" units (targets+snowmen).

I also made it so arrow's can't collide with each other, which fixes a bug where arrows can potentially grow giant due to the increase scale on collision.

Finally, I made it so once arrows collide with something, they are put into their own container.